Remove accelerometer header and set dedicated LED pin
[hw rev. 3]
Since the Hot Potato (game) use case was deemed less significant
hardware support for it was removed in order to leave more space
for the common components. This also enabled us to use a dedicated
pin for the LED which was previously routed in parallel with the
buzzer.  This should enable better control over the hardware and
(theoretically) more current to be supplied to the LED and the
buzzer.